refactor(backend): extract HostedSiteCoordinator from AgentService
PR 2, first commit: the code moves, the test files are untouched and green.

agent-service.ts 4160 -> 3758 (-402); agent/hosted-site-coordinator.ts is 504
lines. The coordinator owns publishing to openbot.site end to end: the approval
a mutation needs, the mutation, and the conversation markers that record it —
`#pendingHostedSiteTerminalEvents`, `#pendingHostedSiteTerminalDeliveries` and
`#hostedSiteTerminalRetryTimer`, plus the sixteen private members that drove
them.

The real change is in respondToApproval: a 48-line inline branch becomes a
four-line delegation to `resolveApproval`. #surfaceHostedSiteApproval keeps only
the parts shared with the other three surfaces — registering the pending
approval, marking the routine as needing attention, emitting — and gets the
approval and mutation from `prepareApproval`.

One deviation from the plan, deliberate. The plan moved
`PendingApproval.hostedSiteMutation?` into a `Map<RequestId,
HostedSiteMutationContext>` inside the coordinator. It stays on the approval
record, typed by the coordinator's exported interface. Request ids come from the
provider process and are reused across client restarts, and three paths drop a
pending approval without going through respondToApproval (provider loss, stop,
turn cleanup), so a second map keyed by request id could hand a stale mutation
to an unrelated approval. PR 4 moves the record with the import intact.

`OpenBotToolResponse` moves to agent/routine-tools.ts, where `openBotToolResult`
already spelled the same shape out as its return type.

Invariant re-checked: `BEGIN IMMEDIATE` still appears exactly once under
src/backend/agent — #appendHostedSiteEvent still goes through
`withConversationTransaction`, and it still runs inside the transaction
respondToApproval may already have open.

Tests: all six agent-service shards plus the two agent/ files, 93 passed. tsc on
tsconfig.node.json clean, which is what proves the public surface is unchanged —
`Pick<AgentService, ...>` in team-api-server.ts names 45 of these methods.
biome clean on agent-service.ts and src/backend/agent/.
